Add type mapping from QAPI 'null' to Java Optional
This also changes the Java API to version 1.8 since we're using Optional
in the library.

As a side effect, the "animal sniffer" check is currently set to skip I
believe because of this bug:
https://github.com/mojohaus/animal-sniffer/pull/32
If org.anarres.gradle:gradle-stdproject-plugin switches to
https://github.com/xvik/gradle-animalsniffer-plugin then animal sniffer
to 1.16 might fix the problem and we can re-enable it.